message publishing

+++ msrest/Stock/Stock.Capabilities/Persistence/States/ProductState.cs
@@ -1,5 +1,5 @@
 namespace Stock.Capabilities.Persistence.States;
 
 public sealed record ProductState(Guid Id, string Name, string Description
-        , float Weight, float Price, byte[] RowVersion)
+        , float Weight, decimal Price, byte[] RowVersion)
     : BaseState(RowVersion);
